+++ BetterCrafting/Managers/FavoriteManager.cs
@@ -28,6 +28,7 @@ public class FavoriteManager : BaseManager {
 	#region Events
 
 	[Subscriber]
+	[EventPriority(EventPriority.Low)]
 	public void OnSaveLoaded(object? sender, SaveLoadedEventArgs e) {
 		LoadFavorites();
 	}
@@ -39,12 +40,31 @@ public class FavoriteManager : BaseManager {
 	// TODO: Store favorites in the save data, which would require
 	// multiplayer packets to sync with remote clients.
 
+	public bool DoesHaveSaveFavorites() {
+		if (string.IsNullOrEmpty(Constants.SaveFolderName))
+			return false;
+
+		Favorites? data;
+		string path = $"savedata/favorites/{Constants.SaveFolderName}.json";
+
+		try {
+			data = Mod.Helper.Data.ReadJsonFile<Favorites>(path);
+		} catch (Exception ex) {
+			Log($"The {path} file is invalid or corrupt.", LogLevel.Error, ex);
+			return false;
+		}
+
+		return data?.Cooking is not null && data?.Crafting is not null;
+	}
+
 	public void LoadFavorites() {
 		if (string.IsNullOrEmpty(Constants.SaveFolderName))
 			return;
 
 		Favorites? data;
-		string path = $"savedata/favorites/{Constants.SaveFolderName}.json";
+		string path = Mod.UseGlobalSave
+			? $"savedata/favorites.json"
+			: $"savedata/favorites/{Constants.SaveFolderName}.json";
 
 		try {
 			data = Mod.Helper.Data.ReadJsonFile<Favorites>(path);
@@ -66,7 +86,9 @@ public class FavoriteManager : BaseManager {
 		if (string.IsNullOrEmpty(Constants.SaveFolderName) || UserFavorites == null || !Modified)
 			return;
 
-		string path = $"savedata/favorites/{Constants.SaveFolderName}.json";
+		string path = Mod.UseGlobalSave
+			? $"savedata/favorites.json"
+			: $"savedata/favorites/{Constants.SaveFolderName}.json";
 
 		try {
 			Mod.Helper.Data.WriteJsonFile(path, UserFavorites);