+++ BetterCrafting/Models/SaveSpecificConfig.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Leclair.Stardew.BetterCrafting.Models;
+
+public class SaveSpecificConfig {
+
+	public bool? UseGlobalSave { get; set; } = null;
+
+}